// client/src/app/payor-tool/payor-tool-data.ts
export class PayorTooldata {
    AllignedDirection: any = 'Left';
    DefaultValue: any = null;
    EquivalentIncomingField: any = null;
    EquivalentLearnedField: any = null;
    EquivalentDeuField: any = null;
    ControlHeight: Number = 23;
    CssProperties: any = null;
    Disabled: boolean = false;
    FieldOrder: any = null;
    ControlX: any = null;
    ControlY: any = null;
    FieldStatusValue: any = null;
    IsRequired: Boolean = true;
    ControlWidth: Number = 200;
    FormulaId: any = null;
    HelpText: any = null;
    IsCalculatedField: Boolean = false;
    IsZeroorBlankAllowed: Boolean = false;
    IsOverrideOfCalcAllowed: Boolean = false;
    IsPopulateIfLinked: Boolean = false;
    IsPartOfPrimaryKey: Boolean = false;
    IsTabbedToNextFieldIfLinked: Boolean = true;
    LabelOnField: any = null;
    MaskFieldTypeId: any = null;
    MaskFieldType: any = null;
    MaskText: any = null;
    MaskFieldList: any = [];
    PTAvailableFieldId: any = null;
    AvailableFieldName: any = null;
    PayorFieldID: any = null;
    PayorToolId: any = null;
}
export class ImportPaymentData {
    ID: Number;
    PayorID: any = undefined;
    TemplateID: any = undefined;
    FieldsID: any = undefined;
    FieldsName: any = undefined;
    Disabled: Boolean = false;
    FixedColLocation: any = undefined;
    FixedRowLocation: any = undefined;
    HeaderSearch: any = undefined;
    RelativeColLocation: any = undefined;
    RelativeRowLocation: any = undefined;
    PayorToolMaskFieldTypeId: any = undefined;
    MaskFieldList: any = [];
    MaskFieldType: any = undefined;
    TransID: any = undefined;
    TransName: any = undefined;
    ImportToolTranslator: any = [];
    strDefaultText: any = undefined;
    PartOfPrimaryKey: Boolean = false;
    CalculatedFields: Boolean = false;
    className: any = undefined;
    FormulaExpression: any = undefined;
}